Refactor authentication-hapi-plugin
