+++ src/authentication/authentication-hapi-plugin.ts
@@ -2,21 +2,17 @@ import * as Boom from 'boom';
 import * as auth from 'hapi-auth-jwt2';
 import { inject, injectable } from 'inversify';
 import * as Knex from 'knex';
-import * as moment from 'moment';
-import * as qs from 'querystring';
 
 import * as Hapi from '../server/hapi';
 import { HapiPlugin } from '../server/hapi-register';
-import { Group, User } from '../shared/gitlab';
+import { Group } from '../shared/gitlab';
 import { GitlabClient } from '../shared/gitlab-client';
+import { Logger, loggerInjectSymbol } from '../shared/logger';
 import { charlesKnexInjectSymbol } from '../shared/types';
+import { validateTeamToken } from './team-token';
 import { AccessToken, jwtOptionsInjectSymbol } from './types';
 
-export interface TeamToken {
-  teamId: number;
-  token: string;
-  createdAt: number | moment.Moment;
-}
+const randomstring = require('randomstring');
 
 export const subEmailClaimKey: 'https://sub_email' = 'https://sub_email';
 export const teamTokenClaimKey: 'https://team_token' = 'https://team_token';
@@ -30,6 +26,7 @@ class AuthenticationHapiPlugin extends HapiPlugin {
     @inject(GitlabClient.injectSymbol) private readonly gitlab: GitlabClient,
     @inject(jwtOptionsInjectSymbol) private readonly hapiOptions: auth.JWTStrategyOptions,
     @inject(charlesKnexInjectSymbol) private readonly db: Knex,
+    @inject(loggerInjectSymbol) private readonly logger: Logger,
   ) {
     super({
       name: 'authentication-plugin',
@@ -56,10 +53,11 @@ class AuthenticationHapiPlugin extends HapiPlugin {
   public async getTeamHandler(request: Hapi.Request, reply: Hapi.IReply) {
     try {
       const credentials = request.auth.credentials as AccessToken;
-      const user = await getUserByEmail(credentials[subEmailClaimKey], this.gitlab);
-      const teams = await getUserTeams(user.id, this.gitlab);
+      const user = await this.gitlab.getUserByEmail(credentials[subEmailClaimKey]);
+      const teams = await this.gitlab.getUserTeams(user.id);
       return reply(teams[0]); // NOTE: we only support a single team for now
     } catch (error) {
+      this.logger.error(`Can't fetch user or team`, error);
       return reply(Boom.wrap(error, 401));
     }
   }
@@ -71,19 +69,23 @@ class AuthenticationHapiPlugin extends HapiPlugin {
       const teamId = await validateTeamToken(teamToken, this.db);
       const team = await this.gitlab.fetchJson<Group>(`groups/${teamId}`);
       const email = credentials[subEmailClaimKey];
+      const password = generatePassword();
       const {id, idp} = parseSub(credentials.sub);
-      const user = await createUser(
+      const user = await this.gitlab.createUser(
         email,
-        '12345678',
+        password,
         email.replace('@', '-'),
         email,
-        this.gitlab,
         id,
         idp,
       );
-      await addUserToGroup(user.id, teamId, this.gitlab);
-      return reply(team);
+      await this.gitlab.addUserToGroup(user.id, teamId);
+      return reply({
+        team,
+        password,
+      });
     } catch (error) {
+      this.logger.error(`Problems on signup`, error);
       return reply(Boom.wrap(error, 401));
     }
   }
@@ -122,90 +124,6 @@ export function parseSub(sub: string) {
   };
 }
 
-export async function getUserByEmail(email: string, gitlab: GitlabClient) {
-  const search = {
-    search: email,
-  };
-  const users = await gitlab.fetchJson<User[]>(`users?${qs.stringify(search)}`, true);
-  if (!users || !users.length) {
-    throw Boom.badRequest(`Can\'t find user '${email}'`);
-  }
-  if (users.length > 1) {
-    // This shoud never happen
-    throw Boom.badRequest(`Found multiple users with email '${email}'`);
-  }
-  return users[0];
-}
-
-export function createUser(
-  email: string,
-  password: string,
-  username: string,
-  name: string,
-  gitlab: GitlabClient,
-  externUid?: string,
-  provider?: string,
-  confirm = false,
-) {
-  const newUser = {
-    email,
-    password,
-    username,
-    name,
-    extern_uid: externUid,
-    provider,
-    confirm,
-  };
-  return gitlab.fetchJson<User>(`users`, {
-    method: 'POST',
-    body: JSON.stringify(newUser),
-    headers: {
-      'content-type': 'application/json',
-    },
-  }, true);
-}
-
-export function addUserToGroup(userId: number, teamId: number, gitlab: GitlabClient, accessLevel = 30) {
-  return gitlab.fetchJson(`groups/${teamId}/members`, {
-    method: 'POST',
-    body: JSON.stringify({
-      id: teamId,
-      user_id: userId,
-      access_level: accessLevel,
-    }),
-    headers: {
-      'content-type': 'application/json',
-    },
-  }, true);
-}
-
-export async function getUserTeams(userId: number, gitlab: GitlabClient) {
-  const sudo = {
-    sudo: userId,
-  };
-  return gitlab.fetchJson<Group[]>(`groups?${qs.stringify(sudo)}`, true);
-}
-
-export function teamTokenQuery(token: string, db: Knex) {
-  if (!token || token.length !== 7 || !token.match(/^\w+$/)) {
-    throw new Error('Invalid team token');
-  }
-  const latestTokens = db('teamtoken')
-    .select(db.raw('teamId, MAX(createdAt) as latestStamp'))
-    .groupBy('teamId')
-    .as('latest');
-  return db('teamtoken')
-    .join(latestTokens, ((join: any) => join
-      .on('teamtoken.teamId', '=', 'latest.teamId')
-      .andOn('teamtoken.createdAt', '=', 'latest.latestStamp')) as any,
-    )
-    .where('teamtoken.token', token);
-}
-
-export async function validateTeamToken(token: string, db: Knex) {
-  const teamTokens: TeamToken[] = await teamTokenQuery(token, db);
-  if (!teamTokens || teamTokens.length !== 1) {
-    throw new Error('Invalid team token');
-  }
-  return teamTokens[0].teamId;
+export function generatePassword(length = 16) {
+  return randomstring.generate({length, charset: 'alphanumeric', readable: true});
 }
